requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
pytest>=6.2.0
pyyaml>=5.4.0
python-dotenv>=0.19.0
tqdm>=4.62.0